:recycle: Switch from using applyJmesPathTransform model method to sub-service pattern.
