--- api/src/models/dataset-integrations/index.ts
@@ -1,1 +1,0 @@
-export { applyJMESPathTransform } from "./apply-jmes-path-transform"
