+++ api/src/services/dataset-integrations/index.ts
@@ -3,5 +3,6 @@ export { UpdateService } from "./update-service"
 
 // Sub-Services
 export { ActivateService } from "./activate-service"
+export { ApplyJMESPathTransformService } from "./apply-jmes-path-transform-service"
 export { BulkReplaceDatasetEntriesService } from "./bulk-replace-dataset-entries-service"
 export { RefreshService } from "./refresh-service"
